* Force load of `BouncyCastleProvider` to ensure `MD4` implementation  exists * Numerous code cleanups, I hope I didn't break anything * Use a `LOGGER` per class instead of using `JISystem.getLogger` * Deprecate `JISystem.setInBuiltLogHandler` * Disable execution of "tests" files when compiling

+++ j-interopdeps/src/main/java/rpc/ConnectionOrientedPdu.java
@@ -211,7 +211,7 @@ public abstract class ConnectionOrientedPdu extends NdrObject implements Protoco
         ndr.setFormat(format);
         setFragmentLength(ndr.readUnsignedShort());
         setAuthenticatorLength(ndr.readUnsignedShort());
-        this.callId = ((int) ndr.readUnsignedLong());
+        this.callId = ndr.readUnsignedLong();
     }
 
     protected void writeHeader(NetworkDataRepresentation ndr) {
